requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==23.2.0
black==24.2.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
dataclasses-json==0.6.4
distro==1.9.0
environs==9.5.0
Flask==3.0.2
frozenlist==1.4.1
greenlet==3.0.3
grpcio==1.60.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.9
langchain-community==0.0.24
langchain-core==0.1.28
langchain-text-splitters==0.0.1
langsmith==0.1.8
lxml==5.1.0
markdown-it-py==3.0.0
markdown2==2.4.13
MarkupSafe==2.1.5
marshmallow==3.20.2
mdurl==0.1.2
minio==7.2.5
multidict==6.0.5
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.26.4
openai==1.13.3
orjson==3.9.15
packaging==23.2
pandas==2.2.1
pathspec==0.12.1
pdfminer.six==20221105
pdfplumber==0.10.4
pillow==10.2.0
platformdirs==4.2.0
protobuf==4.25.3
pyarrow==15.0.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.6.2
pydantic_core==2.16.3
pymilvus==2.3.6
PyPDF2==3.0.1
pypdfium2==4.27.0
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.27
tenacity==8.2.3
text-unidecode==1.3
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
Werkzeug==3.0.1
yarl==1.9.4
